+++ cli/src/commands/mod.rs
@@ -5,6 +5,7 @@ pub mod create;
 pub mod explore;
 pub mod idl;
 pub mod know;
+pub mod programs;
 pub mod public_artifacts;
 pub mod sdk;
 pub mod stack;